change all Ids from int to Long

+++ moviemanager_backend/src/main/java/com/mtcompany/moviemanager/dao/DirectorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DirectorRepository extends JpaRepository<Director,Integer> {
+public interface DirectorRepository extends JpaRepository<Director,Long> {
 
 }
