+++ moviemanager_backend/src/main/java/com/mtcompany/moviemanager/service/ActorService.java
@@ -8,15 +8,15 @@ import java.util.List;
 public interface ActorService {
     List<Actor> findAll();
 
-    Actor findById(int theId);
+    Actor findById(Long theId);
 
     Actor save(Actor theActor);
 
-    void deleteById(int theId);
+    void deleteById(Long theId);
 
-    public List<Movie> getMoviesByActorId(int theId);
+    public List<Movie> getMoviesByActorId(Long theId);
 
-    public Actor addMovie(int actorId,int movieId);
+    public Actor addMovie(Long actorId,Long movieId);
 
-    String removeMovieFromActor(int actorId, int movieId);
+    String removeMovieFromActor(Long actorId, Long movieId);
 }
