+++ moviemanager_backend/src/main/java/com/mtcompany/moviemanager/service/CategoryServiceImpl.java
@@ -23,7 +23,7 @@ public class CategoryServiceImpl implements CategoryService{
     }
 
     @Override
-    public Category findById(int theId) {
+    public Category findById(Long theId) {
         Optional<Category> result = categoryRepository.findById(theId);
 
         Category theCategory = null;
